requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
bitarray==2.9.2
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
cbor2==5.6.4
celery==5.2.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpathlib==0.19.0
confection==0.1.5
crytic-compile==0.3.7
cymem==2.0.8
cytoolz==0.12.3
Deprecated==1.2.13
distro==1.9.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.2
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.4.1
getSourceCode==1.0.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.2.4
langcodes==3.4.0
language_data==1.2.0
lru-dict==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
memory-profiler==0.61.0
multi_key_dict==2.0.3
multidict==6.0.5
murmurhash==1.0.10
numpy==1.26.4
openai==1.35.14
packaging==24.1
parsimonious==0.10.0
pbr==5.10.0
preshed==3.0.9
prettytable==3.10.2
prompt-toolkit==3.0.31
protobuf==5.27.2
psutil==6.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.0.9
python-jenkins==1.7.0
pytz==2022.2.1
pyunormalize==15.1.0
redis==4.3.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.8.1
rlp==4.0.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
slither-analyzer==0.10.3
smart-open==7.0.4
sniffio==1.3.1
solc-select==1.0.4
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssdeep==3.4
thinc==8.2.5
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
vine==5.0.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
web3==6.20.0
websockets==12.0
Werkzeug==2.2.2
wrapt==1.16.0
yarl==1.9.4
z3-solver==4.12.1.0
